added second controller

+++ src/main/java/com/svyat/test/repository/PopulationRepository.java
@@ -3,5 +3,7 @@ package com.svyat.test.repository;
 import com.svyat.test.model.Population;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PopulationRepository extends JpaRepository<Population, Integer> {
+import java.math.BigInteger;
+
+public interface PopulationRepository extends JpaRepository<Population, BigInteger> {
 }
