+++ src/main/java/com/svyat/test/service/PopulationService.java
@@ -2,16 +2,17 @@ package com.svyat.test.service;
 
 import com.svyat.test.model.Population;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface PopulationService {
-    Population getById(Integer id);
+    Population getById(BigInteger id);
 
     void save(Population population);
 
     void saveAll(List<Population> population);
 
-    void delete(Integer id);
+    void delete(BigInteger id);
 
     List<Population> getAll();
 }
